requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain-groq==0.1.0
python-dotenv==1.0.0
pydantic==2.5.2
uvicorn==0.27.0
